refactor: split the recording adapter and the recipe out of the record driver
The record module carried four concerns whose only shared property was the word
record. It now holds the driver alone.

- Moves the recording adapter (RecordedResource, RecordingActivity,
  RecordedDraftBook, RecordingSink, RecordingBookshelf) to publisher/recording.py
  with the private serialisation helpers it uses.
- Moves RecordRecipe and load_record_recipe to publisher/recipe.py.
- Names the composed visibility precedence rule resolve_book_visibility, gives it
  one home in recipe.py, and tests it directly.
- Leaves setup, SetupResult, run_record, parse_parameters and _ACTIVE_RECORDING in
  record.py. The adapter no longer reaches the driver's module state.

The public interface is unchanged. Every name exported from bookshelf.publisher
imports from the same path and behaves the same.
